requirements.txt
flask
mysql-connector-python
opencv-python-headless
numpy
pillow
pytesseract
matplotlib
gunicorn
python-dotenv
